Fix AGP version: 8.10.1 does not exist, use 8.10.0
AGP 8.10.1 was never published to Google Maven. Change to 8.10.0
which is the latest confirmed stable release in the 8.10.x series
and is compatible with Kotlin 2.1.21 and Gradle 8.11.1.

Also fix Int-to-Float subtraction in drawText topLeft calculation.

+++ app/src/main/java/com/chordquiz/app/ui/components/chord/ChordDiagram.kt
@@ -95,7 +95,7 @@ fun ChordDiagramCanvas(
                 textLayoutResult = fretTextLayout,
                 topLeft = Offset(
                     leftPad - size.width * 0.12f,
-                    topPad + fretSpacing * 0.6f - fretTextLayout.size.height
+                    topPad + fretSpacing * 0.6f - fretTextLayout.size.height.toFloat()
                 )
             )
         }
